Outlet control UI fully implemented
The Outlet control UI is fully implemented.  This involves calls to
the BluetoothManager to fetch/update outlets.

At the moment the 'fetch outlets' action populates the UI with a
hard-coded set of outlets and the 'update outlet' action is a no-op.
As soon as the communication protocol is implemented these commands
will be relayed to the SmartOutlet network coordinator.